+++ packages/code-du-travail-frontend/src/outils/ConventionCollective/agreement/SearchAgreementInput.tsx
@@ -1,17 +1,20 @@
 import {
   Input,
   Label,
+  Paragraph,
   Section as SectionUi,
   Text,
   theme,
 } from "@socialgouv/cdtn-ui";
 import React, { ForwardedRef } from "react";
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 
 import { InfoBulle } from "../../common/InfoBulle";
 import { SectionTitle } from "../../common/stepStyles";
+import EmbeddedInForm from "../common/EmbeddedInForm";
 
 type Props = {
+  embeddedForm: boolean;
   query?: string;
   onChange: (event: React.ChangeEvent) => void;
   placeholder?: string;
@@ -21,19 +24,36 @@ const defaultPlaceholder = "Ex : Transports routiers ou 1486";
 
 export const SearchAgreementInput = React.forwardRef(
   function _SearchAgreementInput(
-    { query = "", onChange, placeholder = defaultPlaceholder }: Props,
+    {
+      embeddedForm = true,
+      query = "",
+      onChange,
+      placeholder = defaultPlaceholder,
+    }: Props,
     ref: ForwardedRef<HTMLFormElement>
   ): JSX.Element {
     return (
       <Section>
-        <SectionTitle>
-          Précisez et sélectionnez votre convention collective
-        </SectionTitle>
-        <form ref={ref} onSubmit={(event) => event.preventDefault()}>
-          <InlineLabel htmlFor="agreement-search">
+        {embeddedForm ? (
+          <SectionTitle>
+            Précisez et sélectionnez votre convention collective
+          </SectionTitle>
+        ) : (
+          <ParagraphNoMarginBottom fontWeight="600" fontSize="default">
+            Précisez et sélectionnez votre convention collective
+          </ParagraphNoMarginBottom>
+        )}
+        <EmbeddedInForm
+          enable={embeddedForm}
+          reference={ref}
+          onSubmit={(event) => event.preventDefault()}
+        >
+          <InlineLabel htmlFor="agreement-search" embeddedForm={embeddedForm}>
             Nom de la convention collective ou son numéro d’identification{" "}
             <abbr title="Identifiant de la Convention Collective">IDCC</abbr>{" "}
-            <Text fontWeight="400">(champ obligatoire)</Text>
+            <Text fontWeight="400" fontSize="small">
+              (champ obligatoire)
+            </Text>
           </InlineLabel>
           <InfoBulle title="Qu'est ce qu'un IDCC">
             <p>
@@ -58,22 +78,30 @@ export const SearchAgreementInput = React.forwardRef(
             onChange={onChange}
             autoComplete="off"
           />
-        </form>
+        </EmbeddedInForm>
       </Section>
     );
   }
 );
-
 const Section = styled(SectionUi)`
   padding-top: 0;
 `;
+const ParagraphNoMarginBottom = styled(Paragraph)`
+  margin-bottom: 0;
+`;
 
 const BlockInput = styled(Input)`
-  padding-top: ${theme.spacings.base};
+  padding-top: ${theme.spacings.small};
   width: 100%;
 `;
 
 const InlineLabel = styled(Label)`
   display: inline;
-  margin-right: 0;
+  ${(props) => {
+    if (!props.embeddedForm) {
+      return css`
+        font-weight: 400;
+      `;
+    }
+  }}
 `;