+++ packages/code-du-travail-frontend/src/outils/common/Agreement/RouteSelection/index.ts
@@ -1,0 +1,3 @@
+import RouteSelection from "./RouteSelection";
+
+export { RouteSelection };